Add little doc-comments and split some files
