+++ lib/data_access_layer/data_sources/data_source.dart
@@ -3,8 +3,10 @@
 import 'package:meta/meta.dart';
 import 'package:tao_cat/data_access_layer/document_model.dart';
 
+/// An abstract interface for data sources that load and save documents.
 abstract interface class DataSource {
   @useResult
   Future<DocumentModel> loadDocument();
+
   Future<void> saveDocument(DocumentModel document);
 }
