+++ lib/data_access_layer/data_sources/random_document_data_source.dart
@@ -7,6 +7,7 @@ import 'package:tao_cat/data_access_layer/segment_model.dart';
 
 import 'package:flutter/foundation.dart';
 
+/// A data source for generating random documents.
 @immutable
 class RandomDocumentDataSource implements DataSource {
   RandomDocumentDataSource({
