+++ lib/data_access_layer/segment_model.dart
@@ -1,9 +1,13 @@
+/// A model representing a segment within a document.
 class SegmentModel {
   SegmentModel({
     required this.sourceText,
     required this.isTitle,
   });
 
+  /// The source text of the segment.
   String sourceText;
+
+  /// Indicates if the segment is a title.
   bool isTitle;
 }
