+++ lib/my_business_layer/document_bloc.dart
@@ -3,14 +3,17 @@ import 'package:tao_cat/domain_layer/document_service.dart';
 import 'package:tao_cat/my_business_layer/document_event.dart';
 import 'package:tao_cat/my_business_layer/document_state.dart';
 
+/// A bloc that handles events and states related to document operations.
 class DocumentBloc extends Bloc<DocumentEvent, DocumentState> {
   DocumentBloc(this.documentService) : super(DocumentInitial()) {
+    /// Handles the event of loading a document.
     on<LoadDocumentEvent>((event, emit) async {
       emit(DocumentLoading());
       final document = await documentService.loadDocument(event.source);
       emit(DocumentLoaded(document));
     });
 
+    /// Handles the event of editing a document.
     on<EditDocumentEvent>((event, emit) async {
       if (state case DocumentLoaded(:final document)) {
         final updatedDocument = await documentService.editDocument(
@@ -22,13 +25,14 @@ class DocumentBloc extends Bloc<DocumentEvent, DocumentState> {
       }
     });
 
+    /// Handles the event of saving a document.
     on<SaveDocumentEvent>((event, emit) async {
       if (state case DocumentLoaded(:final document)) {
-        print('on<SaveDocumentEvent>');
         await documentService.saveDocument(document);
       }
     });
   }
 
+  /// The service that performs document-related operations.
   final DocumentService documentService;
 }
